Made changes to the LED strip connection
Due to a short circuit caused by the LED strip and power supply onto the bread board we changed the hardware. We implemented a terminal block on a PCB board that connects to the +/- terminals of the board board. The LED strip, power supply, and terminal block are connected through a 3 way splice.

Overall resulting in changes to the BOM, Images, and READ.ME
Note: Keyed Connecter usage changed to terminal block and JST connector use